refactor(agent): extract revision.py from graph.py
