+++ app/agent/graph.py
@@ -18,26 +18,23 @@ import logging
 from typing import Any, Literal
 
 from langchain_core.language_models import BaseChatModel
-from langchain_core.messages import AIMessage, BaseMessage, HumanMessage, SystemMessage, ToolMessage
+from langchain_core.messages import AIMessage, BaseMessage, SystemMessage, ToolMessage
 from langgraph.graph import END, StateGraph
 from pydantic import BaseModel
 
 from app.agent.commit import commit_stops
-from app.agent.critique import (
-    CRITIQUE_ITINERARY,
-    CRITIQUE_STEP,
-    CRITIQUE_VIBE,
-    vibe,
-)
-from app.agent.critique.checks import itinerary_violations
+from app.agent.critique import vibe
 from app.agent.prompts import SYSTEM_PROMPT
-from app.agent.state import ItineraryState, RevisionHint, Stop
+from app.agent.revision import (
+    critique_final_with_stops,
+    critique_step,
+    finalize_as_is,
+    short_circuit_max_steps,
+)
+from app.agent.state import ItineraryState, Stop
 from app.agent.tools import COMMIT_ITINERARY_TOOL_NAME, all_tools
 
 logger = logging.getLogger(__name__)
-
-LOW_SIMILARITY_THRESHOLD = 0.55
-MAX_REVISIONS_PER_REASON = 2
 
 
 _RECENT_TOOL_EXCHANGES_KEPT = 2
@@ -98,256 +95,6 @@ def _serialize_tool_result(result: Any) -> str:
     if isinstance(result, list) and result and isinstance(result[0], BaseModel):
         return json.dumps([m.model_dump(mode="json") for m in result])
     return json.dumps(result, default=str)
-
-
-def _can_retry(state: ItineraryState, reason: str) -> bool:
-    return state.revision_counts.get(reason, 0) < MAX_REVISIONS_PER_REASON
-
-
-def _bumped_counts(state: ItineraryState, reason: str) -> dict[str, int]:
-    counts = dict(state.revision_counts)
-    counts[reason] = counts.get(reason, 0) + 1
-    return counts
-
-
-def _scratch_entries_for_last_round(
-    state: ItineraryState,
-) -> list[tuple[str, dict[str, Any]]]:
-    """Return (tool_name, scratch_entry) pairs for every tool_call in the most
-    recent issuing AIMessage, in tool_call order.
-
-    Pairings are matched by tool_call id when present, falling back to the
-    highest-step entry for that tool name (handles legacy entries that
-    pre-date id tracking)."""
-    last_tool_idx = None
-    for i in range(len(state.messages) - 1, -1, -1):
-        if isinstance(state.messages[i], ToolMessage):
-            last_tool_idx = i
-            break
-    if last_tool_idx is None:
-        return []
-    issuing_ai: AIMessage | None = None
-    for i in range(last_tool_idx - 1, -1, -1):
-        m = state.messages[i]
-        if isinstance(m, AIMessage) and m.tool_calls:
-            issuing_ai = m
-            break
-    if issuing_ai is None or not issuing_ai.tool_calls:
-        return []
-    pairs: list[tuple[str, dict[str, Any]]] = []
-    for tc in issuing_ai.tool_calls:
-        name = tc["name"]
-        entries = state.scratch.get(name) or []
-        if not entries:
-            continue
-        match = next((e for e in entries if e.get("id") == tc["id"]), None)
-        if match is None:
-            match = max(entries, key=lambda e: e.get("step", -1))
-        pairs.append((name, match))
-    return pairs
-
-
-def _most_restrictive_filter(filters: dict[str, Any] | None) -> str:
-    """Deterministic priority order for which filter to drop first."""
-    if not filters:
-        return "none"
-    for f in ("open_at", "price_level_max", "min_rating", "neighborhood", "types_any"):
-        if filters.get(f) is not None:
-            return f
-    return "none"
-
-
-def _diagnose_one(tool_name: str, entry: dict[str, Any]) -> RevisionHint | None:
-    """Inspect a single tool call+result pair. Returns a hint or None."""
-    result = entry.get("result")
-    args = entry.get("args") or {}
-
-    if isinstance(result, dict) and "error" in result:
-        return RevisionHint(
-            reason="tool_error",
-            detail=str(result["error"]),
-            suggested_action="try_different_tool",
-            target={"tool": tool_name},
-        )
-
-    if isinstance(result, list):
-        if not result:
-            return RevisionHint(
-                reason="empty_results",
-                detail=f"No matches for {args}.",
-                suggested_action="drop_filter",
-                target={"filter": _most_restrictive_filter(args.get("filters"))},
-            )
-        if all(getattr(h, "business_status", None) != "OPERATIONAL" for h in result):
-            return RevisionHint(
-                reason="all_closed",
-                detail="Every result is closed or permanently_closed.",
-                suggested_action="broaden_query",
-                target={"filter": "business_status"},
-            )
-        top = result[0]
-        sim = getattr(top, "similarity", 0.0) or 0.0
-        if sim < LOW_SIMILARITY_THRESHOLD:
-            return RevisionHint(
-                reason="low_similarity",
-                detail=f"Top similarity {sim:.2f} below threshold {LOW_SIMILARITY_THRESHOLD}.",
-                suggested_action="broaden_query",
-                target={"query": args.get("query")},
-            )
-    return None
-
-
-def _diagnose_last_tool_result(state: ItineraryState) -> RevisionHint | None:
-    """Diagnose every tool_call in the most recent issuing AIMessage and
-    return the first hint in tool_call order. Returns None if every call was
-    healthy. The agent revises one issue per round; later issues, if any,
-    will be diagnosed on the next round."""
-    for tool_name, entry in _scratch_entries_for_last_round(state):
-        hint = _diagnose_one(tool_name, entry)
-        if hint is not None:
-            return hint
-    return None
-
-
-def _hint_for_violation(reason: str, state: ItineraryState) -> RevisionHint:
-    """Map a check name from itinerary_violations() to a structured hint."""
-    if reason == "geographic_coherence":
-        return RevisionHint(
-            reason="geographic_incoherence",
-            detail="One or more consecutive stops exceed the per-leg walking budget.",
-            suggested_action="tighten_radius",
-            target={"stops": list(range(1, len(state.stops)))},
-        )
-    if reason == "temporal_coherence":
-        return RevisionHint(
-            reason="temporal_incoherence",
-            detail="A stop is closed at its planned arrival time.",
-            suggested_action="shift_arrival_time",
-            target={},
-        )
-    if reason == "walking_budget_respected":
-        return RevisionHint(
-            reason="walking_budget_exceeded",
-            detail="Total walking exceeds the user's budget.",
-            suggested_action="rebalance_walking_budget",
-            target={},
-        )
-    if reason == "no_hallucinated_place_ids":
-        return RevisionHint(
-            reason="hallucinated_place_id",
-            detail="One or more place_ids do not exist in places_raw.",
-            suggested_action="swap_stop",
-            target={},
-        )
-    return RevisionHint(
-        reason="constraint_unmet_in_final",
-        detail="Final stops do not satisfy the shared user constraints.",
-        suggested_action="swap_stop",
-        target={},
-    )
-
-
-def _final_with_caveats(last_content: str, violations: list[str]) -> str:
-    """Compose a final reply that lists what didn't quite work. Better than
-    silently shipping a bad plan."""
-    caveats = (
-        "\n\nCaveats: I couldn't fully satisfy "
-        + ", ".join(violations)
-        + (" after revisions. You may want to adjust the plan.")
-    )
-    return (last_content or "") + caveats
-
-
-def _last_ai_content(last: BaseMessage | None) -> str:
-    return last.content if isinstance(last, AIMessage) and isinstance(last.content, str) else ""
-
-
-def _short_circuit_max_steps(state: ItineraryState) -> dict[str, Any]:
-    return {
-        "done": True,
-        "final_reply": state.final_reply
-        or "I hit the planning step limit. Here is the best plan I had so far.",
-    }
-
-
-def _finalize_as_is(state: ItineraryState, last: BaseMessage | None) -> dict[str, Any]:
-    return {"done": True, "final_reply": state.final_reply or _last_ai_content(last)}
-
-
-def _critique_final_with_stops(
-    state: ItineraryState,
-    last: BaseMessage | None,
-    judge_llm: BaseChatModel | None,
-) -> dict[str, Any]:
-    """Run deterministic checks; if any fail, drive a revision (or ship with
-    caveats once budgets are exhausted). If they all pass, run the optional
-    cheap-LLM vibe check; if that fails, drive a revision. Otherwise finalize."""
-    last_content = _last_ai_content(last)
-    violations = itinerary_violations(state)
-    if violations:
-        actionable = next((v for v in violations if _can_retry(state, v)), None)
-        if actionable is not None:
-            hint = _hint_for_violation(actionable, state)
-            return {
-                "revision_hints": [*state.revision_hints, hint],
-                "revision_counts": _bumped_counts(state, actionable),
-                "messages": [
-                    HumanMessage(
-                        content=(
-                            f"{CRITIQUE_ITINERARY} {actionable}: {hint.detail} "
-                            f"Suggested action: {hint.suggested_action}. "
-                            f"Revise the affected stop(s) and re-call commit_itinerary."
-                        )
-                    )
-                ],
-                "done": False,
-            }
-        return {"done": True, "final_reply": _final_with_caveats(last_content, violations)}
-
-    score = vibe.vibe_check(state, judge_llm)
-    if score is not None and score < vibe.VIBE_THRESHOLD and _can_retry(state, "vibe_mismatch"):
-        hint = RevisionHint(
-            reason="vibe_mismatch",
-            detail=f"Cross-stop vibe coherence scored {score:.1f}/5.",
-            suggested_action="swap_stop",
-            target={},
-        )
-        return {
-            "revision_hints": [*state.revision_hints, hint],
-            "revision_counts": _bumped_counts(state, "vibe_mismatch"),
-            "messages": [
-                HumanMessage(
-                    content=(
-                        f"{CRITIQUE_VIBE} cross-stop vibe coherence scored "
-                        f"{score:.1f}/5. Swap whichever stop feels off and "
-                        f"re-call commit_itinerary."
-                    )
-                )
-            ],
-            "done": False,
-        }
-
-    return {"done": True, "final_reply": state.final_reply or last_content}
-
-
-def _critique_step(state: ItineraryState) -> dict[str, Any]:
-    """React to the last tool result. Emit a per-step hint if it was empty,
-    all-closed, low-similarity, or errored. Bounded by retry budget."""
-    hint = _diagnose_last_tool_result(state)
-    if hint is None or not _can_retry(state, hint.reason):
-        return {}
-    return {
-        "revision_hints": [*state.revision_hints, hint],
-        "revision_counts": _bumped_counts(state, hint.reason),
-        "messages": [
-            HumanMessage(
-                content=(
-                    f"{CRITIQUE_STEP} {hint.reason}: {hint.detail} "
-                    f"Suggested next action: {hint.suggested_action} on {hint.target}."
-                )
-            )
-        ],
-    }
 
 
 def build_agent_graph(
@@ -460,16 +207,16 @@ def build_agent_graph(
         per-step diagnose. Each branch is a pure function returning the
         LangGraph update dict."""
         if state.step_count >= max_steps:
-            return _short_circuit_max_steps(state)
+            return short_circuit_max_steps(state)
 
         last = state.messages[-1] if state.messages else None
         finalizing = isinstance(last, AIMessage) and not last.tool_calls
 
         if finalizing and state.stops:
-            return _critique_final_with_stops(state, last, judge_llm)
+            return critique_final_with_stops(state, last, judge_llm)
         if finalizing:
-            return _finalize_as_is(state, last)
-        return _critique_step(state)
+            return finalize_as_is(state, last)
+        return critique_step(state)
 
     def route_after_plan(state: ItineraryState) -> Literal["act", "critique"]:
         last = state.messages[-1]
